pyproject.toml
[project]
name = "vai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "baml-py>=0.214.0",
    "duckdb>=1.4.2",
    "fastapi>=0.123.5",
    "google-genai>=1.52.0",
    "groq>=0.37.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pydub>=0.25.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
